pipeline/requirements.txt
anthropic>=0.26.0
openai>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0
python-dateutil>=2.9.0
requests>=2.32.0

# Test dependencies
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-mock>=3.12.0
responses>=0.25.0
